Make GPU disk post-start optional

+++ .devcontainer/post_start.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
+set -e
 
-if [ ! -f /tmp/x86-ubuntu-gpu-ml-isca ]; then
+if [ -f x86-ubuntu-gpu-ml-isca.gz ] && [ ! -f /tmp/x86-ubuntu-gpu-ml-isca ]; then
   cp x86-ubuntu-gpu-ml-isca.gz /tmp
   gunzip /tmp/x86-ubuntu-gpu-ml-isca.gz
 fi
 
-ln -s /tmp/x86-ubuntu-gpu-ml-isca .
+if [ -f /tmp/x86-ubuntu-gpu-ml-isca ] && [ ! -e x86-ubuntu-gpu-ml-isca ]; then
+  ln -s /tmp/x86-ubuntu-gpu-ml-isca .
+fi
